int to size_t

+++ libsrc/mat3x3.cpp
@@ -516,14 +516,14 @@ mat3x3 mat3x3_covariance(std::vector<vec3> &points)
 
 	vec3 mean = make_vec3(0, 0, 0);
 
-	for (int i = 0; i < points.size(); i++)
+	for (size_t i = 0; i < points.size(); i++)
 	{
 		mean = vec3_add_vec3(mean, points[i]);
 	}
 
 	vec3_mult(&mean, 1 / (double)points.size());
 
-	for (int k = 0; k < points.size(); k++)
+	for (size_t k = 0; k < points.size(); k++)
 	{
 		for (int j = 0; j < 3; j++)
 		{
